Converted the webpack configuration file from CommonJS to ES Module. Fixed DllReferencePlugin require() issue. Handled CKEditorTranslationsPlugin correctly and Updated NonCKEPluginEntries() to return an object instead of using map().
